Reused DiningDetail
